Rename firmare module.
