+++ tests/test_ota_firmware.py
@@ -2,7 +2,7 @@ from unittest import mock
 
 import pytest
 
-import zigpy.ota.firmware as firmware
+import zigpy.ota.image as firmware
 import zigpy.types as t
 
 MANUFACTURER_ID = mock.sentinel.manufacturer_id
