+++ zigpy/ota/image.py
